set NIX_LDFLAGS="-lgcc_s" to make it available for pthread_cancel
Running mdrun with threads can call pthread_cancel, and the
gcc-wrapper does not make libgcc_s.so.1 available.  One solution was
to set the LD_LIBRARY_PATH, but that wasn't really tenable.

It turns out this is a common issue:
http://permalink.gmane.org/gmane.linux.distributions.nixos/11610

My understanding of the accepted solution is to add -lgcc_s to NIX_LDFALGS:
http://permalink.gmane.org/gmane.linux.distributions.nixos/11611